Initialize the state before rendering widgets editor
